Do not overwrite existing files on unpack
Interestingly rootfs.img file is in archive that I checked twice. First
time it is indeed rootfs that seems to be boring, then it is raw
entropy. So looks like we want the first one :) And I would check the
contents with more attention.

+++ unpack_apn.py
@@ -24,6 +24,9 @@ for entry in archive.contents:
   parentdir = fullpath.parent
   if not parentdir.exists():
     parentdir.mkdir(parents=True)
+  if fullpath.exists():
+    print(f'Refusing to overwrite {fullpath}', file=sys.stderr)
+    continue
   with open(fullpath, 'wb') as fp:
     fp.write(entry.contents)
   print(entry.filename)
